TaijaData

Documentation for TaijaData.

TaijaData.data_catalogue

TaijaData.load_blobs

TaijaData.load_california_housing

TaijaData.load_cifar_10

TaijaData.load_cifar_10_test

TaijaData.load_circles

TaijaData.load_credit_default

TaijaData.load_fashion_mnist

TaijaData.load_fashion_mnist_test

TaijaData.load_german_credit

TaijaData.load_gmsc

TaijaData.load_linearly_separable

TaijaData.load_mnist

TaijaData.load_mnist_test

TaijaData.load_moons

TaijaData.load_multi_class

TaijaData.load_overlapping

TaijaData.load_synthetic_data

TaijaData.load_tabular_data

TaijaData.load_uci_adult

TaijaData.data_catalogue — Constant

A dictionary that provides an overview of the various benchmark datasets and the methods to load them.

TaijaData.load_blobs — Function

load_blobs(n=250; seed=data_seed, kwrgs...)

Loads overlapping synthetic data.

TaijaData.load_california_housing — Function

load_california_housing(n::Union{Nothing,Int}=5000)

Loads California Housing data.

TaijaData.load_cifar_10 — Function

load_cifar_10(n::Union{Nothing, Int}=nothing)

Loads data from the CIFAR-10 dataset.

TaijaData.load_cifar_10_test — Method

load_cifar_10_test()

Loads test data from the CIFAR-10 dataset.

TaijaData.load_circles — Function

load_circles(n=250; seed=data_seed, noise=0.15, factor=0.01)

Loads synthetic circles data.

TaijaData.load_credit_default — Function

load_credit_default(n::Union{Nothing,Int}=5000)

Loads UCI Credit Default data.

TaijaData.load_fashion_mnist — Function

load_fashion_mnist(n::Union{Nothing,Int}=nothing)

Loads FashionMNIST data.

TaijaData.load_fashion_mnist_test — Method

load_fashion_mnist_test()

Loads FashionMNIST test data.

TaijaData.load_german_credit — Function

load_german_credit(n::Union{Nothing, Int}=nothing)

Loads UCI German Credit data.

TaijaData.load_gmsc — Function

load_gmsc(n::Union{Nothing,Int}=5000)

Loads Give Me Some Credit (GMSC) data.

TaijaData.load_linearly_separable — Function

load_linearly_separable(n=250; seed=data_seed)

Loads linearly separable synthetic data.

TaijaData.load_mnist — Function

load_mnist(n::Union{Nothing,Int}=nothing)

Loads MNIST data.

TaijaData.load_mnist_test — Method

load_mnist_test()

Loads MNIST test data.

TaijaData.load_moons — Function

load_moons(n=250; seed=data_seed, kwrgs...)

Loads synthetic moons data.

TaijaData.load_multi_class — Function

load_multi_class(n=250; seed=data_seed)

Loads multi-class synthetic data.

TaijaData.load_overlapping — Function

load_overlapping(n=250; seed=data_seed)

Loads overlapping synthetic data.

TaijaData.load_synthetic_data — Function

load_synthetic_data(n=250; seed=data_seed)

Loads all synthetic datasets and wraps them in a dictionary.

TaijaData.load_tabular_data — Function

load_tabular_data(n=nothing; drop=nothing)

Loads all tabular datasets and wraps them in a dictionary.

TaijaData.load_uci_adult — Function

load_uci_adult(n::Union{Nothing, Int}=1000)

Loads data from the UCI 'Adult' dataset.